Update README.md
The Python notebooks and slides in this repository describe the methods I used for building classifiers using logistic regression and random forest to link infant death records with their corresponding birth records.  In the past, this process was undertaken either manually or using SQL queries.  The goal of this project was to see if supervised machine learning classifiers could improve on these previous methods in terms of effort and accuracy.