Update README.md
The Python notebooks in this repository document the methods used to build two supervised machine learning classifiers to link infant death records with their corresponding birth records.  Previously, this linkage was completed either manually or using SQL queries.  The goal of this project was to explore whether logistic regression and random forest classifiers can improve on the prior methods in terms of effort and accuracy.